[WIP] cleaning up MAC accessing. Invalid mailbox request.
