+++ system/platforms/arm-rpi3/Makerules
@@ -29,6 +29,7 @@ C_FILES = setupStack.c       \
 		  unparkcore.c		 \
 		  mmu.c				 \
 		  random.c			 \
+		  bcm2837_mbox.c		 \
 		  des.c
 
 # Add the files to the compile source path
